test: normalize test filenames to <source>_test.py convention
Rename test files that did not follow the <source>_test.py pattern and
merge those whose target name was already taken into the existing module
test file:

Renames (content-preserving):
- _dense/plasticity_test.py        -> plasticity_binary_test.py
- _op/kernix_runtime_load_test.py  -> kernix_runtime_test.py
- pararnn/_fused_ad_test.py        -> _fused_test.py

Merges (appended to existing <source>_test.py, sources removed):
- _csr/mm_golden_parity_test.py    -> _csr/main_test.py
- _fcn/golden_parity_test.py       -> _fcn/main_test.py
- _fcn/mm_golden_parity_test.py    -> _fcn/main_test.py (deduped shared _cases)

Net test count unchanged (whole-project collection: 2547). Merged tests
inherit each target module's slow marker, so they now run in the full
lane. Feature-suite files (kernix_*) and _fcn/conversion_test.py are
intentionally left as-is.
