+++ brainevent/_csr/main_test.py
@@ -25,6 +25,7 @@ import numpy as np
 import pytest
 
 import brainevent
+from brainevent import CSR, CSC, BinaryArray
 from brainevent._csr.test_util import get_csr, vector_csr, matrix_csr, csr_vector, csr_matrix
 
 # Every test in this module dispatches to the native ``numba`` backend (the default backend),
@@ -1408,3 +1409,101 @@ def test_csc_update_on_post_method_parity():
     ref = _be.update_csc_on_binary_post(
         csc.data, csc.indices, csc.indptr, pre_trace, post_spike, 0.0, 1.5, shape=csc.shape)
     assert _jnp.allclose(got, ref, atol=1e-6)
+
+
+# --- merged from brainevent/_csr/mm_golden_parity_test.py ---
+# Golden parity for binary CSR/CSC matrix-matrix dispatch against ``M.todense()``.
+
+
+def _csr_cases():
+    rng = np.random.default_rng(0)
+    m, k = 6, 5
+    for homo in (True, False):
+        dense = (rng.random((m, k)) + 0.5) * (rng.random((m, k)) > 0.5)
+        dense = jnp.asarray(dense, dtype=jnp.float32)
+        csr = CSR.fromdense(dense)
+        if homo:
+            csr = CSR((jnp.asarray([1.5], jnp.float32), csr.indices, csr.indptr), shape=csr.shape)
+        yield csr
+
+
+def _mask(x, dtype):
+    x = jnp.asarray(x)
+    return jnp.asarray(x > 0, dtype=dtype) if x.dtype != jnp.bool_ else jnp.asarray(x, dtype=dtype)
+
+
+def test_csr_matmat_golden():
+    rng = np.random.default_rng(7)
+    for csr in _csr_cases():
+        dense = jnp.asarray(csr.todense(), dtype=jnp.float32)
+        m, k = csr.shape
+        for ev in (jnp.bool_, jnp.float32):
+            for n in (1, 3, 8):
+                right = jnp.asarray(rng.random((k, n)) > 0.5, dtype=ev)   # CSR @ M (unfavorable)
+                got_r = csr @ BinaryArray(right)
+                ref_r = dense @ _mask(right, jnp.float32)
+                assert jnp.allclose(got_r, ref_r, atol=1e-5), ('CSR@M', str(ev), n)
+
+                left = jnp.asarray(rng.random((n, m)) > 0.5, dtype=ev)    # M @ CSR (favorable)
+                got_l = BinaryArray(left) @ csr
+                ref_l = _mask(left, jnp.float32) @ dense
+                assert jnp.allclose(got_l, ref_l, atol=1e-5), ('M@CSR', str(ev), n)
+
+
+def test_csc_matmat_golden():
+    rng = np.random.default_rng(11)
+    for csr in _csr_cases():
+        csc = csr.T                          # CSC view; dense transposes too
+        dense = jnp.asarray(csc.todense(), dtype=jnp.float32)
+        p, q = csc.shape
+        for ev in (jnp.bool_, jnp.float32):
+            for n in (1, 3, 8):
+                right = jnp.asarray(rng.random((q, n)) > 0.5, dtype=ev)   # CSC @ M (favorable)
+                got_r = csc @ BinaryArray(right)
+                ref_r = dense @ _mask(right, jnp.float32)
+                assert jnp.allclose(got_r, ref_r, atol=1e-5), ('CSC@M', str(ev), n)
+
+                left = jnp.asarray(rng.random((n, p)) > 0.5, dtype=ev)    # M @ CSC (unfavorable)
+                got_l = BinaryArray(left) @ csc
+                ref_l = _mask(left, jnp.float32) @ dense
+                assert jnp.allclose(got_l, ref_l, atol=1e-5), ('M@CSC', str(ev), n)
+
+
+def test_csr_matmat_unfavorable_builds_weight_indices():
+    # Routing to the indexed scatter primitive populates the cached transposed
+    # structure as a side effect (the gather path never builds it).
+    # CSR._weight_indices() caches the CSC view under the 'csc' buffer key.
+    rng = np.random.default_rng(2)
+    csr = next(_csr_cases())
+    k = csr.shape[1]
+    right = jnp.asarray(rng.random((k, 4)) > 0.5, dtype=jnp.bool_)
+    assert csr.buffers.get('csc') is None
+    _ = csr @ BinaryArray(right)
+    assert csr.buffers.get('csc') is not None
+
+
+def test_csc_matmat_unfavorable_builds_weight_indices():
+    # CSC._weight_indices() caches the CSR view under the 'csr' buffer key.
+    rng = np.random.default_rng(3)
+    csc = next(_csr_cases()).T
+    p = csc.shape[0]
+    left = jnp.asarray(rng.random((4, p)) > 0.5, dtype=jnp.bool_)
+    assert csc.buffers.get('csr') is None
+    _ = BinaryArray(left) @ csc
+    assert csc.buffers.get('csr') is not None
+
+
+def test_csr_matmat_units_and_jit():
+    rng = np.random.default_rng(9)
+    csr = next(_csr_cases())
+    m, k = csr.shape
+    csr_u = CSR((csr.data * u.mV, csr.indices, csr.indptr), shape=csr.shape)
+    M = jnp.asarray(rng.random((k, 4)) > 0.5, dtype=jnp.bool_)
+    got = csr_u @ BinaryArray(M)
+    assert u.get_unit(got) == u.mV
+    ref = (jnp.asarray(csr.todense(), jnp.float32) @ jnp.asarray(M, jnp.float32))
+    assert jnp.allclose(u.get_mantissa(got), ref, atol=1e-5)
+
+    f = jax.jit(lambda d: CSR((d, csr.indices, csr.indptr), shape=csr.shape) @ BinaryArray(M))
+    got_jit = f(csr.data)
+    assert jnp.allclose(got_jit, ref, atol=1e-5)
